+++ src/js/features/sessions/state/creator/initialState.js
@@ -1,0 +1,11 @@
+export default {
+  request: {
+    comment: '',
+    distanceMiles: 0.0,
+    machineType: '',
+    seconds: 0,
+    startDate: new Date(),
+    timing: '00:00:00',
+    userName: ''
+  }
+};