+++ src/js/features/sessions/state/reducer.js
@@ -1,16 +1,16 @@
 import initialState from './initialState';
-import * as types from './types';
+import * as _types from './types';
 
 const reducer = (state = initialState, action) => {
   let newState;
   switch (action.type) {
-    case types.INVALIDATED_CARDIO_MACHINE_SESSIONS:
+    case _types.INVALIDATED_CARDIO_MACHINE_SESSIONS:
       newState = { ...initialState, didInvalidate: true, isFetching: false };
       break;
-    case types.REQUEST_CARDIO_MACHINE_SESSIONS:
+    case _types.REQUEST_CARDIO_MACHINE_SESSIONS:
       newState = { ...state, didInvalidate: false, isFetching: true };
       break;
-    case types.RECIEVE_CARDIO_MACHINE_SESSIONS:
+    case _types.RECIEVE_CARDIO_MACHINE_SESSIONS:
       newState = {
         ...state,
         currentPage: action.metaPayload._currentPage,
@@ -24,7 +24,7 @@ const reducer = (state = initialState, action) => {
       };
       break;
     default:
-      newState = state;
+      newState = { ...state };
   }
   return newState;
 };