+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6.0] - 2019-11-02
+### Added
+- Prop validation for new props coming into session creator class
+
+### Changed
+- Moved rountine list request into redux store
+- Moved caffeine, exercise, and session create post logic into redux
 
 ## [0.5.0] - 2019-11-01
 ### Changed